move ClusterOverview utils to the new overview comp
